Refactor: Datastar messages for fragments
